# blueprint/openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele Wedding Registry API
  description: >
    Pool-funded wedding gift registry platform with kinship-tier gift
    suggestions, multi-currency support, and ACID-safe contributions.
  version: "1.0.0"
  contact:
    name: Saukele Support
    email: support@saukele.kz
servers:
  - url: http://localhost:3000/api
    description: Local development server
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token. Obtain via POST /auth/login.
        Include as: Authorization: Bearer <token>
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        code:
          type: string
          nullable: true
          example: VALIDATION_ERROR
        error:
          type: string
          example: Resource not found
        message:
          type: string
          nullable: true
          example: Validation failed
        details:
          type: object
          nullable: true
        requestId:
          type: string
          nullable: true

    # ── Monetary precision-safe scalars ───────────────────────────────────
...
    CursorMeta:
      type: object
      properties:
        nextCursor:
          type: integer
          nullable: true
          example: 42
          description: Pass as ?cursor= in next request. Null means no more pages.
        take:
          type: integer
          example: 20
        hasMore:
          type: boolean
          example: true

    # ── User ──────────────────────────────────────────────────────────────
...
    FamilyMemberCreate:
      type: object
      required: [name, relationshipType, kinshipTier]
      properties:
        relatedUserId:
          type: integer
          nullable: true
          example: 8
          description: If relative is a registered user; null if not
        name:
          type: string
          example: Нұрлан Қасымов (Ата)
        relationshipType:
          type: string
          enum: [PARENT, SIBLING, CHILD, COUSIN, UNCLE_AUNT, FAMILY_FRIEND, OTHER]
          example: PARENT
        kinshipTier:
          type: string
          enum: [CLOSE, EXTENDED, FRIEND]
          example: CLOSE
    FamilyMemberResponse:
      type: object
      properties:
        id:
          type: integer
          example: 4
        userId:
          type: integer
          example: 1
        relatedUserId:
          type: integer
          nullable: true
          example: 8
        name:
          type: string
          example: Нұрлан Қасымов (Ата)
        relationshipType:
          type: string
          example: PARENT
        kinshipTier:
          type: string
          example: CLOSE
        createdAt:
          type: string
          format: date-time
    KinshipTierResponse:
      type: object
      properties:
        kinshipTier:
          type: string
          enum: [CLOSE, EXTENDED, FRIEND]
          nullable: true
          example: EXTENDED
        suggestedMinKzt:
          type: integer
          example: 20000
...
paths:
...
  /family-members:
    post:
      tags: [FamilyMembers]
      summary: Add a family member to kinship tree (REGISTRANT or ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FamilyMemberCreate'
      responses:
        '201':
          description: Family member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyMemberResponse'
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Role must be REGISTRANT or ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid enum value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags: [FamilyMembers]
      summary: List own family members (ADMIN can list any via query filters)
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
          description: ADMIN only. List family tree entries for a specific user.
        - name: cursor
          in: query
          schema:
            type: integer
        - name: take
          in: query
          schema:
            type: integer
            default: 50
        - name: kinshipTier
          in: query
          schema:
            type: string
            enum: [CLOSE, EXTENDED, FRIEND]
      responses:
        '200':
          description: Paginated family member list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FamilyMemberResponse'
                  meta:
                    $ref: '#/components/schemas/CursorMeta'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /family-members/{id}:
    delete:
      tags: [FamilyMembers]
      summary: Remove a family member entry (owner or ADMIN)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Family member removed
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not the owner of this entry and not ADMIN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Family member entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /family-members/kinship-tier/{userId}:
    get:
      tags: [FamilyMembers]
      summary: Determine kinship tier of a given user relative to the caller
      description: >
        Uses recursive CTE to traverse the family tree.
        Returns kinshipTier and suggested minimum contribution in KZT.
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 8
      responses:
        '200':
          description: Kinship tier result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KinshipTierResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── ADMIN ─────────────────────────────────────────────────────────────────
...